+++ src/imagetextalternative/imagetextalternativeediting.js
@@ -4,7 +4,7 @@
  */
 
 /**
- * @module image/imagetextalternative/imagetextalternativeengine
+ * @module image/imagetextalternative/imagetextalternativeediting
  */
 
 import ImageTextAlternativeCommand from './imagetextalternativecommand';
@@ -16,7 +16,7 @@ import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
  *
  * @extends module:core/plugin~Plugin
  */
-export default class ImageTextAlternativeEngine extends Plugin {
+export default class ImageTextAlternativeEditing extends Plugin {
 	/**
 	 * @inheritDoc
 	 */
