+++ tests/imagetextalternative/imagetextalternativeediting.js
@@ -4,22 +4,23 @@
  */
 
 import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor';
-import ImageTextAlternativeEngine from '../../src/imagetextalternative/imagetextalternativeengine';
+import ImageTextAlternativeEditing from '../../src/imagetextalternative/imagetextalternativeediting';
 import ImageTextAlternativeCommand from '../../src/imagetextalternative/imagetextalternativecommand';
 
-describe( 'ImageTextAlternativeEngine', () => {
+describe( 'ImageTextAlternativeEditing', () => {
 	let editor;
+
 	beforeEach( () => {
 		return VirtualTestEditor
 			.create( {
-				plugins: [ ImageTextAlternativeEngine ]
+				plugins: [ ImageTextAlternativeEditing ]
 			} )
 			.then( newEditor => {
 				editor = newEditor;
 			} );
 	} );
 
-	it( 'should register ImageAlteranteTextCommand', () => {
+	it( 'should register ImageAlternativeTextCommand', () => {
 		expect( editor.commands.get( 'imageTextAlternative' ) ).to.be.instanceOf( ImageTextAlternativeCommand );
 	} );
 } );
